Implemented DbController with dapper
