+++ CodingTracker.S1m0n32002/CodingTracker.S1m0n32002.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35828.75 d17.13
+VisualStudioVersion = 17.13.35828.75
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodingTracker.S1m0n32002", "CodingTracker.S1m0n32002\CodingTracker.S1m0n32002.csproj", "{4FE92486-87AE-4265-99E4-6B0FB9F3E0CE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpectreExplorer", "..\..\..\repo\Utils\SpectreExplorer\SpectreExplorer\SpectreExplorer.csproj", "{984FA7C8-1575-4842-91DA-F005912F1B8A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{4FE92486-87AE-4265-99E4-6B0FB9F3E0CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4FE92486-87AE-4265-99E4-6B0FB9F3E0CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4FE92486-87AE-4265-99E4-6B0FB9F3E0CE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{984FA7C8-1575-4842-91DA-F005912F1B8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{984FA7C8-1575-4842-91DA-F005912F1B8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{984FA7C8-1575-4842-91DA-F005912F1B8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{984FA7C8-1575-4842-91DA-F005912F1B8A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
