+++ CodingTracker.S1m0n32002/CodingTracker.S1m0n32002/Program.cs
@@ -1,13 +1,9 @@
 ﻿using CodingTracker.S1m0n32002.Models;
-using System.Diagnostics;
+using Dapper;
+using System.Data;
+using System.Data.SQLite;
 
-class Program
-{
-    static void Main(string[] args)
-    {
-        Settings.Load();
-        Settings.Save();
-        Debug.WriteLine($"DbName: {Settings.Current.DbName}");
-        Debug.WriteLine($"DbPath: {Settings.Current.DbPath}");
-    }
-}
+Settings.Load();
+
+var dbController = new CodingTracker.S1m0n32002.Controllers.DbController();
+
