+++ CodingTracker.S1m0n32002/CodingTracker.S1m0n32002/Settings/AppSettings.json
@@ -1,4 +1,4 @@
 {
-  "DbName": "CodingTracker.db",
-  "DbPath": "Database"
+  "Db": "Database\\Database.sqlite",
+  "Template": "Templates\\SqliteTemplate.sql"
 }
