+++ CodingTracker.S1m0n32002/CodingTracker.S1m0n32002/Templates/SqliteTemplate.sql
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS "Sessions" (
+	"Id"	INTEGER NOT NULL UNIQUE,
+	"Description"	TEXT,
+	"Start"	TEXT NOT NULL,
+	"End"	TEXT,
+	PRIMARY KEY("Id")
+);
