--- CodingTracker.S1m0n32002/CodingTracker.S1m0n32002/Templates/SqliteTemplate.txt
@@ -1,13 +1,0 @@
-CREATE TABLE IF NOT EXISTS "habits" (
-	"id"	INTEGER NOT NULL UNIQUE,
-	"name"	TEXT NOT NULL UNIQUE COLLATE NOCASE,
-	"periodicity"	INTEGER NOT NULL,
-	PRIMARY KEY("id")
-);
-CREATE TABLE IF NOT EXISTS "occurrences" (
-	"id"	INTEGER,
-	"date"	TEXT,
-	"habitid"	INTEGER,
-	PRIMARY KEY("id"),
-	FOREIGN KEY("habitid") REFERENCES "habits"("id")
-);
